requirements.txt
pyinstaller
pynput
requests
pillow
flask